Created let-them-eat-cake: boop coldpost timeWhat's your favorite dessert?
It's alright if you don't know. The Foundation offers you a lot of choices. One of the nice things about working here is how well our needs are met. Hell, might as well stop saying we work here. It's a lot more than that. It's a community, a lifestyle, we are the folks behind the curtain making sure the trains run on time and that time can run the trains. Easy to be proud of the work we do here.

Getting back on the topic of desserts. We've got a few notoriously dangerous snacks. Chief among them would be our friends the Keter Cakes. I'm sure you're familiar. We don't fill our bellies then all other space will become cake. Maybe you've worked on it and know the lingo. Dumpling-Class personnel waddling in to a top secret mission. That's another one of the funny things that happens when you've got full buy-in to the Foundation's way of doing things. Dehumanizing outsiders becomes second nature.

Quick question- do you think we serve the greater good of humanity, or normalcy? Trick question. Same thing. At least they are as far as the Foundation is concerned. Another similarity is that they're nebulous and vague concepts. 

Warren G. Harding popularized the phrase 'return to normalcy' as a political slogan. One of those thing where it sounds nice to the ear without actually meaning a thing. 

If you ever really dig into the file on the Keter Cakes, they supposedly just popped into existence one day and then just kept happening. That's awful convenient. A lot of faith to put in the Foundation to keep it under wraps? A death wish? No. Those pastries are not as they appear.

The Keter Cakes have a unique property. They increase in correlation to the hunger of humanity. Which is absurd, right? Isn't it a little inconvenient to have exponentially increasing cakes that slowly, inefficiently, crawl their way towards wherever the hungry and dying lay in the streets? Hunger isn't going to be solved that way.

You might even be one of the sharp cookies out there who knows that hunger is a distribution problem and not a supply-side issue. Ever since Fritz Haber gave us that good fertilizer we've had no problem keeping production on par with the amount of hungry mouths. So isn't it just as ridiculous that there's still grief and death from empty bellies?

If you've ever seen the pictures from Syria, Yemen, Congo, the Holocaust, Holodomor, and a million other things then you've got a taste of how horrible hunger is. There's a lot of different actors to blame. It's always the fault of humanity. We choose to feed some and let others starve. That's normalcy.

What sort of normalcy are we preserving? The Foundation has infinitely replicating food and the infrastructure to instantly transport it anyplace where a child's belly is dry and shrinking while their limbs waste into thin rods and their faces age a hundred years. It's normal for us to grind the bones of the dead into bread so that we may live to see more die another day.

There's no proof to what I know the Cakes do. Does it matter? We have the resources, means, and opportunity to preserve a normalcy where everyone can eat enough to live. What sort of protection are we offering a world where the population of a small town in Ohio is extinguished every day because the abundant food resources available to everyone else didn't get to them in time? 

We will make a desert, and call it peace.